# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WEB MÌNH",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lấy danh sách danh mục",
                "description": "API trả về danh sách danh mục. Có thể lấy theo cấu trúc cây hoặc danh sách phẳng. Mặc định chỉ lấy danh mục đang hoạt động.",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "parameters": [
                    {
                        "name": "tree",
                        "in": "query",
                        "description": "Nếu bằng 1, trả về cấu trúc cây (lồng nhau)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Nếu bằng 1, lấy cả danh mục bị ẩn (Admin dùng)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Tạo danh mục mới (Admin)",
                "operationId": "bb02f0320882db37f99574196577e446",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Tên danh mục",
                                        "type": "string",
                                        "example": "Danh mục mới"
                                    },
                                    "parent_id": {
                                        "description": "ID danh mục cha (để trống hoặc 0 nếu là gốc)",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "image": {
                                        "description": "File ảnh danh mục",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "Mô tả ngắn",
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "description": "Trạng thái (1: Hiện, 0: Ẩn)",
                                        "type": "integer",
                                        "default": 1,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "sort_order": {
                                        "description": "Thứ tự hiển thị (càng nhỏ càng ưu tiên)",
                                        "type": "integer",
                                        "default": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Chi tiết danh mục",
                "operationId": "4cdf41756315109f1efc72787db5830a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Cập nhật danh mục (Admin)",
                "description": "Vì Laravel không hỗ trợ multipart/form-data qua PUT tốt, nên dùng POST kèm _method=PUT",
                "operationId": "8ea50e48cbcd07043a3fd8145e99bfbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "sort_order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Xóa danh mục (Admin)",
                "operationId": "f1b8af6e84bd52b9594d98bde748ebd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Không thể xóa vì còn sản phẩm"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Category",
                "description": "Mô hình Danh mục sản phẩm",
                "properties": {
                    "id": {
                        "description": "ID duy nhất của danh mục",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Tên hiển thị của danh mục",
                        "type": "string",
                        "example": "Nội thất phòng khách"
                    },
                    "slug": {
                        "description": "Đường dẫn thân thiện (tự động tạo từ tên + unique id)",
                        "type": "string",
                        "example": "noi-that-phong-khach-65e8a1"
                    },
                    "parent_id": {
                        "description": "ID của danh mục cha (null nếu là danh mục gốc)",
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "image": {
                        "description": "Đường dẫn ảnh danh mục (lưu trong storage)",
                        "type": "string",
                        "example": "categories/image.jpg",
                        "nullable": true
                    },
                    "description": {
                        "description": "Mô tả chi tiết về danh mục",
                        "type": "string",
                        "example": "Mô tả danh mục",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Trạng thái hiển thị (true: hiện, false: ẩn)",
                        "type": "boolean",
                        "example": true
                    },
                    "sort_order": {
                        "description": "Thứ tự sắp xếp (số càng nhỏ càng hiện lên đầu)",
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "description": "Ngày tạo",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Ngày cập nhật gần nhất",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}